[JSContext] Define `ContextExecutor.setContextName` for debugging
Add a method that lets JS set the name of the JSContext for debugging purposes. I check `JSGlobalContextSetName` since it is not available on iOS 7.

Test Plan: Run `NativeModules.ContextExecutor.setContextName('x')` with the default executor active, and find it from the Safari develop menu with the correct name. Changed the name when the debugger was connected and saw that its name was updated the next time I opened the menu.